refactor(help): add new directive for plot descriptions
new directive formats the description of the plots in the
menu.It has multiple acces points that can be used to get the descrition
to the right place. Also a new Factory was added that can be used to
match those acces ids with the acces points.
Old descriptions have been replaced now by the new directive.
When a new plot is added this directive has to be used to make its
descripotion!